+++ src/Calendar/Calendar.js
@@ -7,6 +7,7 @@ import moment from 'moment';
 import PropTypes from 'prop-types';
 import { isDateBetween, isEnabledDate } from '../utils/dateUtils';
 import React, { Component } from 'react';
+import 'fundamental-styles/dist/calendar.css';
 
 /** A **Calendar** is commonly used as the contents of a **Popover** when composing a **DatePicker**.
 It is rarely used on its own as a standalone component. */
@@ -48,9 +49,6 @@ class Calendar extends Component {
     }
 
     componentDidMount = () => {
-        if (!this.props.disableStyles) {
-            require('fundamental-styles/dist/calendar.css');
-        }
         this.gridManager = new GridManager(this.getGridOptions());
     }
 
@@ -465,7 +463,6 @@ class Calendar extends Component {
                         <Button
                             aria-label={previousButtonLabel}
                             compact={compact}
-                            disableStyles={this.props.disableStyles}
                             glyph='slim-arrow-left'
                             onClick={this.handlePrevious}
                             option='transparent' />
@@ -473,7 +470,6 @@ class Calendar extends Component {
                     <div className='fd-calendar__action'>
                         <Button
                             compact={compact}
-                            disableStyles={this.props.disableStyles}
                             onClick={this.showMonths}
                             option='transparent'>
                             <span>
@@ -484,7 +480,6 @@ class Calendar extends Component {
                     <div className='fd-calendar__action'>
                         <Button
                             compact={compact}
-                            disableStyles={this.props.disableStyles}
                             onClick={this.showYears}
                             option='transparent'>
                             <span>
@@ -497,7 +492,6 @@ class Calendar extends Component {
                         <Button
                             aria-label={nextButtonLabel}
                             compact={compact}
-                            disableStyles={this.props.disableStyles}
                             glyph='slim-arrow-right'
                             onClick={this.handleNext}
                             option='transparent' />
@@ -506,7 +500,6 @@ class Calendar extends Component {
                         <div className='fd-calendar__action'>
                             <Button
                                 compact={compact}
-                                disableStyles={this.props.disableStyles}
                                 onClick={this.handleToday}
                                 option={'transparent'}>
                                 {this.props.localizedText.todayLabel}
@@ -662,7 +655,6 @@ class Calendar extends Component {
         const {
             compact,
             enableRangeSelection,
-            disableStyles,
             disableWeekends,
             disableBeforeDate,
             disableAfterDate,
@@ -737,8 +729,6 @@ Calendar.propTypes = {
     disableFutureDates: PropTypes.bool,
     /** Set to **true** to disable dates before today\'s date */
     disablePastDates: PropTypes.bool,
-    /** Internal use only */
-    disableStyles: PropTypes.bool,
     /** Disables dates that match a weekday. For example, `disableWeekday={[\'Tuesday\', \'Thursday\', \'Friday\']}` */
     disableWeekday: PropTypes.arrayOf(PropTypes.string),
     /** Set to **true** to disables dates that match a weekend */