+++ src/Forms/FormGroup.js
@@ -1,15 +1,10 @@
 import PropTypes from 'prop-types';
-import React, { useEffect } from 'react';
+import React from 'react';
+import 'fundamental-styles/dist/form-group.css';
 
 /** A **FormGroup** encapsulates components related to a single form input.
  * At a minimum, it should contain **FormLabel** and input child components. */
-const FormGroup = React.forwardRef(({ children, disableStyles, ...props }, ref) => {
-
-    useEffect(() => {
-        if (!disableStyles) {
-            require('fundamental-styles/dist/form-group.css');
-        }
-    }, []);
+const FormGroup = React.forwardRef(({ children, ...props }, ref) => {
 
     // Note that we don't want to include the fd-form-group--inline
     // class here because it is only for FormRadioGroup
@@ -22,9 +17,7 @@ FormGroup.displayName = 'FormGroup';
 
 FormGroup.propTypes = {
     /** Node(s) to render within the component */
-    children: PropTypes.node,
-    /** Internal use only */
-    disableStyles: PropTypes.bool
+    children: PropTypes.node
 };
 
 export default FormGroup;