+++ src/List/List.js
@@ -7,7 +7,8 @@ import ListItem from './_ListItem';
 import ListSelection from './_ListSelection';
 import ListText from './_ListText';
 import PropTypes from 'prop-types';
-import React, { useEffect } from 'react';
+import React from 'react';
+import 'fundamental-styles/dist/list.css';
 
 /** **List** and **Table** are similar as both usually contain a vertical list of data,
 but lists generally contain basic data and tables tend to hold more complex data.
@@ -16,19 +17,12 @@ const List = React.forwardRef(({
     children,
     className,
     compact,
-    disableStyles,
     hasByline,
     navigation,
     noBorder,
     partialNavigation,
     ...props
 }, ref) => {
-
-    useEffect(() => {
-        if (!disableStyles) {
-            require('fundamental-styles/dist/list.css');
-        }
-    }, []);
 
     const ListClasses = classnames(
         'fd-list',
@@ -59,8 +53,6 @@ List.propTypes = {
     className: PropTypes.string,
     /** Set to **true** to enable compact mode */
     compact: PropTypes.bool,
-    /** Internal use only */
-    disableStyles: PropTypes.bool,
     /** Set to **true** if any list item has a byline. */
     hasByline: PropTypes.bool,
     /** Set to **true** if all list items are links */